Change the icon for "Improve this page"
Changed the icon for "Improve this page" to the symbol of two pens. All the icons used here are released under the SIL Open Font License (OFL) or the GNU General Public License (GPL) so they are open in some ways. This commit is related to the issue https://github.com/okfn/opendatahandbook/issues/158.