Update events.md
We are not central, we are Open Knowledge International. 

Also, we don't use meetup anymore. Time to kill it! 
We should add OK Festival as an example of OK international... (just a line)

Also, try to use OK - for a network member, or OKI to refer to Open Knowledge Centeral or OKFN